Move script entry point
