+++ archinstall/__main__.py
@@ -1,4 +1,4 @@
-import archinstall
+from archinstall.main import main
 
 if __name__ == '__main__':
-	archinstall.run_as_a_module()
+	main()
