+++ archinstall/scripts/only_hd.py
@@ -1,13 +1,13 @@
 import sys
 from pathlib import Path
 
-from archinstall import debug, error
 from archinstall.lib.args import arch_config_handler
 from archinstall.lib.configuration import ConfigurationOutput
 from archinstall.lib.disk.filesystem import FilesystemHandler
 from archinstall.lib.disk.utils import disk_layouts
 from archinstall.lib.global_menu import GlobalMenu
 from archinstall.lib.installer import Installer
+from archinstall.lib.output import debug, error
 
 
 def ask_user_questions() -> None:
